Payment To Farmer Added

+++ src/main/java/com/dairy/dto/paymentToFarmer/PaymentToFarmerRequestDto.java
@@ -8,8 +8,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,17 +18,13 @@ public class PaymentToFarmerRequestDto {
 	private Date from_date;
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private Date to_date;
-	private Long amount;
-
+	private Float amount;
 	private Long farmer;
-
 	private int branch;
-	
 	private Long total_collected_milk;
 	private String milktype;
-	
-	private Long feed_deduction;
-	private Long advance_deduction;
+	private Float feed_deduction;
+	private Float advance_deduction;
 	private String payment_method;
 	private String payment_note;
 }
